refactor: 컨트롤러 복잡도 개선
