--- src/main/java/chess/presentation/controller/Controller.java
@@ -1,6 +1,0 @@
-package chess.presentation.controller;
-
-public interface Controller {
-
-    void execute();
-}
