+++ src/main/java/chess/presentation/view/OutputView.java
@@ -17,8 +17,10 @@ public class OutputView {
         System.out.println("> 게임 이동 : move source위치 target위치 - 예. move b2 b3");
     }
 
-    public void printGameEndMessage() {
-        System.out.println("체스 게임을 종료합니다.");
+    public void printContinueGuide() {
+        System.out.println("> 이전에 진행하던 게임이 존재합니다. 불러오겠습니까?");
+        System.out.println("> 이전 계속 진행 : yes");
+        System.out.println("> 새로운 게임 시작 : no");
     }
 
     public void printNewGameMessage() {
@@ -27,6 +29,10 @@ public class OutputView {
 
     public void printContinueMessage() {
         System.out.println("이전에 진행하던 게임을 불러왔습니다.");
+    }
+
+    public void printGameEndMessage() {
+        System.out.println("체스 게임을 종료합니다.");
     }
 
     public void printPieces(Pieces pieces) {
